migrate to frb 2.1

See merge request wallet/proton-wallet-app!242

+++ lib/managers/providers/server.transaction.data.provider.dart
@@ -174,8 +174,8 @@ class ServerTransactionDataProvider extends DataProvider {
         sign: "",
         // TODO:: add sign once apiClient update for it
         exchangeRateTime: walletTransaction.exchangeRate!.exchangeRateTime,
-        exchangeRate: walletTransaction.exchangeRate!.exchangeRate,
-        cents: walletTransaction.exchangeRate!.cents,
+        exchangeRate: walletTransaction.exchangeRate!.exchangeRate.toInt(),
+        cents: walletTransaction.exchangeRate!.cents.toInt(),
       );
       await exchangeRateDao.insert(exchangeRateModel);
     }
